transport_ocr_system/services/ocr-service/requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# OCR и ML
paddlepaddle==2.6.0
paddleocr==2.7.3
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
timm==0.9.16

# LayoutLM
datasets==2.16.1
accelerate==0.25.1

# Preprocessing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scipy==1.11.4
numpy==1.26.3

# MinIO (S3 совместимое хранилище)
minio==7.2.0

# Асинхронные задачи
celery==5.3.6
redis==5.0.1

# Логирование и мониторинг
python-json-logger==2.0.7
prometheus-client==0.19.0
structlog==24.1.0

# Утилиты
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0
aiofiles==23.2.1
